Added TAP protocol
Implemented TAP backend, renamed 'output' to 'pretty' and moved
examples to become test, using TAP with meson. Also removed the behaviour where
one failing test would cause the process to return an EXIT_FAILURE, as it is a valid
assumption to say that the test driver is not failing, only tests are. Still, now allows
the fini hook to modify the process return status, to allow test benches to modify this behaviour at will.
* examples: Moved to test/, executed with the TAP protocol.
* examples/finit.c: Removed, merged with the assert test to make a more complete example.
* src/libcover/backends/tap.c: TAP protocol backend (version <13).
* src/libcover/suite.c: Removed EXIT_FAILURE on test failure, allows fini call to modify cover's return status.